libkbfs: Divide tlf_handle.go into tlf_handle{,_resolve}.go
